+++ views/amendment/_view_text.php
@@ -16,12 +16,21 @@ if ($amendment->hasAlternativeProposaltext() && (
         ($amendment->proposalFeedbackHasBeenRequested() && $amendment->iAmInitiator())
     )) {
     $hasProposedChange = true;
-    $reference         = $amendment->proposalReference;
+    $reference         = $amendment->getAlternativeProposaltextReference();
     if ($reference) {
+        /** @var Amendment $referenceAmendment */
+        $referenceAmendment = $reference['amendment'];
+        /** @var Amendment $reference */
+        $reference = $reference['modification'];
+
         /** @var AmendmentSection[] $sections */
-        $sections = $amendment->proposalReference->getSortedSections(false);
+        $sections = $reference->getSortedSections(false);
         foreach ($sections as $section) {
-            $prefix = \Yii::t('amend', 'proposed_procedure_title');
+            if ($referenceAmendment->id === $amendment->id) {
+                $prefix = \Yii::t('amend', 'pprocedure_title_own');
+            } else {
+                $prefix = \Yii::t('amend', 'pprocedure_title_other') . ' ' . $referenceAmendment->titlePrefix;
+            }
             if (!$amendment->isProposalPublic()) {
                 $prefix = '[ADMIN] ' . $prefix;
             }
@@ -33,7 +42,7 @@ if ($amendment->hasAlternativeProposaltext() && (
 }
 
 
-if ($amendment->changeEditorial != '') {
+if ($amendment->changeEditorial !== '') {
     echo '<section id="section_editorial" class="motionTextHolder">';
     echo '<h3 class="green">' . \Yii::t('amend', 'editorial_hint') . '</h3>';
     echo '<div class="paragraph"><div class="text">';
@@ -49,7 +58,7 @@ foreach ($sections as $section) {
 }
 
 
-if ($amendment->changeExplanation != '') {
+if ($amendment->changeExplanation !== '') {
     echo '<section id="amendmentExplanation" class="motionTextHolder">';
     echo '<h3 class="green">' . \Yii::t('amend', 'reason') . '</h3>';
     echo '<div class="paragraph"><div class="text">';