+++ History.md
@@ -18,6 +18,7 @@
   - Proposed procedures are now an optional functionality per motion type. They are disabled by default.
   - When exporting the proposed procedure list into ODS, there is now an option to also include the comments.
   - When sending a notification to users regarding a proposed procedure, the content of the mail can now be modified by the person sending the notification.
+  - If an amendment is obsoleted by another amendment with a proposed modification, this modification is also shown in the context of the (proposed) obsoleted amendment.
   - Bugfix: It was not possible to delete admin comments in the proposed procedure.
   - Buffix: Prevent a bug when creating a proposed procedure that collides with another amendment.
 - When creating a motion, the confirmation page now shows a preview of the generated PDF.